+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * Implement job profiles list. UIDEXP-80.
 * Accommodate UI to the changes in progress field structure for job execution. UIDEXP-106.
 * Mapping profiles list - integation with backend. UIDEXP-57.
+* Add job profiles form. UIDEXP-82.
 
 ## [1.0.2](https://github.com/folio-org/ui-data-export/tree/v1.0.2) (2020-04-03)
 [Full Changelog](https://github.com/folio-org/ui-data-export/tree/v1.0.1...v1.0.2)